requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fire==0.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
PyJWT==2.10.0
PyYAML==6.0.2
rav==0.0.9
rich==13.9.4
sqlparse==0.5.2
termcolor==2.5.0
tzdata==2024.2
